Server listening update:
While building the docker configs, the server is listening only on localhost (127.0.0.1 and ::1), which means it's only accepting connections from within the container itself.

We need to be able to communicate with the dashboard container, which is on the outside. By adding the host property to the listen call, we can listen on all network interfaces (0.0.0.0) instead of just localhost.

Co-authored-by: Sean Mentele 56010359+SeanM0124@users.noreply.github.com Co-authored-by: Gautam Panakkal 111669555+gpanakkal@users.noreply.github.com